Added dog condition radials
Added icon assets
added dog condition radials (water, food, energy)
made it a prefab  in a scene that is called when the start button is pressed.

Radials will charge to preset percentage. currently that is 100%. the energy meter will fill up faster. this is all customisable an accessible from the scripts on the radials.