Tweaked stuff and added treats
Radials now do not decay as fast.
Radials now save only twice a second instead of 60 times a second.
Adjustes prices of items in shop to reflect the values they give when used.
Added treats to the shop (and made them expensive in comparison) so when a person isnt well or when it is evening and the dog has low energy, you can give him a treat.
Added debug scene where you can manage variables used in the game, here you can also set doglevels to target level and add items. Also added option to do so directly into the dogradials but you need to enable this in the dogradials themselves as well. this is on purpose, but can be used for more things.